Fix for share channel message moderating

+++ modules/misc/share_check.js
@@ -3,14 +3,14 @@ const path = require('path');
 module.exports = async (message, client, Discord) => {
     // YouTube channel
     if (message?.channel.id === process.env.YOUTUBE_CHAN) {
-        if (!message?.content.includes('youtube.com/') || !message?.content.includes('youtu.be/')) {
+        if (!message?.content.includes('youtube.com/') && !message?.content.includes('youtu.be/')) {
             message?.delete().catch(err => console.error(`${path.basename(__filename)} There was a problem deleting a message: `, err));
         }
     }
 
     // Twitch channel
     if (message?.channel.id === process.env.TWITCH_CHAN) {
-        if (!message?.content.includes('twitch.com/') || !message?.content.includes('twitch.tv/')) {
+        if (!message?.content.includes('twitch.com/') && !message?.content.includes('twitch.tv/')) {
             message?.delete().catch(err => console.error(`${path.basename(__filename)} There was a problem deleting a message: `, err));
         }
     }
@@ -31,14 +31,14 @@ module.exports = async (message, client, Discord) => {
 
     // Twitter channel
     if (message?.channel.id === process.env.TWITTER_CHAN) {
-        if (!message?.content.includes('twitter.com/') || !message?.content.includes('t.co/')) {
+        if (!message?.content.includes('twitter.com/') & !message?.content.includes('t.co/')) {
             message?.delete().catch(err => console.error(`${path.basename(__filename)} There was a problem deleting a message: `, err));
         }
     }
 
     // Spotify channel
     if (message?.channel.id === process.env.SPOTIFY_CHAN) {
-        if (!message?.content.includes('spotify.com/') || !message?.content.includes('soundcloud.com')) {
+        if (!message?.content.includes('spotify.com/') && !message?.content.includes('soundcloud.com')) {
             message?.delete().catch(err => console.error(`${path.basename(__filename)} There was a problem deleting a message: `, err));
         }
     }
